feat: Add delete functionality to EventDetails component
The code changes in `EventDetails.jsx` add the ability to delete an event. This is achieved by using the `useMutation` hook from `@tanstack/react-query` to handle the deletion mutation. The `handleDelete` function triggers the mutation, which calls the `deleteEvent` function from `http.js` and updates the `queryClient` to invalidate the "events" query. After a successful deletion, the component navigates to the "/events" page. This new functionality improves the user experience by allowing events to be easily deleted.