refactor: Update deleteEvent mutation in EventDetails component
The code changes in `EventDetails.jsx` refactor the `deleteEvent` mutation to use the `useMutation` hook from `@tanstack/react-query`. The `queryClient.invalidateQueries` call is modified to include the `refetchType` option set to "none". This improves the performance of the deletion process and ensures that the "events" query is not automatically refetched after a successful deletion. The component also navigates to the "/events" page.