refactor: Update fetchEvents function to include max parameter
The code changes in `http.js` modify the `fetchEvents` function to include a `max` parameter. This allows the function to fetch a specified maximum number of events from the backend API. The `max` parameter is used to construct the query string for the API request. This refactor enhances the flexibility and customization of event fetching in the application.